requirements.txt
numpy
Flask
tensorflow-cpu
gunicorn
keras
pillow
